Switch class and config names
